[refactor]: dto/input & dto/result 로 분리
