+++ src/main/java/lotto/view/InputView.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
-import lotto.dto.PurchaseMoneyDto;
-import lotto.dto.WinningConditionDto;
+import lotto.dto.input.PurchaseMoneyDto;
+import lotto.dto.input.WinningConditionDto;
 import lotto.view.printable.QuestionPrintable;
 
 import java.util.Scanner;
